main: drop requirement that main package be part of module
Prior to being called gobin, this program was call goget. The semantics
implied by the "get" part of the name made this wrong, and also made it
hard to justify supporting main packages that were _not_ part of a
module, because their dependencies would potentially be a forever moving
feast.

When we switched to gobin, we added a -g flag to make that "get" step
explicit (with the exception of the initial "get" which is required).
Hence dependencies would only "move" when a -g flag is provided, which
could hardly be described as a surprise to the caller given that they
have supplied a flag described as "get".

Furthermore, it seems far more useful to have gobin work for
non-module-based main packages for now. So this is probably a win all
round.

Fixes #11