main: support installing github.com/myitcv/gobin (#17)
This is initial support that is unlikely to work on Windows, hence the
TODO in the code.

Unfortunately, I've had to mark the test as skip for now because we
can't install "self" from the module under test. I.e. it's not possible
to run go install for a package in the module under test, i.e.
github.com/myitcv/gobin.