main: handle GO111MODULE being set to off in the gobin environment. (#33)
If the caller of gobin as GO111MODULE=off set for whatever reason, we
were previously passing this value through to the go commands that gobin
calls. Given we do everything in a module-aware mode, this is not the
correct behaviour. Hence we need to explicitly set GO111MODULE=on.