main: ensure replace directives are applied for non-'-m' installs. (#48)
We were previously, erroneously, not applying replace directives from
the main package's module's go.mod file for non-'-m' installs. When
calling gobin with -m, it is expected that such replace directives are
not applied. But without -m, we effectively treat the main package's
module as the main module (except that we actually treat it as a
dependency, in order to capture the version information in the resulting
binary).

This commit fixes that discrepancy, and locks it in with a two tests,
that compare the presence or otherwise of the replace directive between
the non-'-m' and -m modes.