deps: temporarily remove requirement on myitcv.io
Because of https://github.com/golang/go/issues/29773, circular module
requirements create a problem with Go 1.11 and 1.12beta2 (at least).

The fix for this is scheduled for Go 1.13, but we need to add a
dependency from myitcv.io -> github.com/myitcv/gobin now.

Given that the dependency on myitcv.io is only a development-time tool
dependency, the easiest thing to do for now is to drop that requirement,
thereby breaking the cycle.