main: fix -m mode stale binary bug (#84)
In 6038506 we incorrectly optimised away an install step when -m is
supplied. We must always install if we are in -m mode, because the main
module is responsible for resolving versions. We therefore utilise the
install step to effectively ensure the target binary is up to date.

This logic would change if were to adopt #81 because we would then only
install non-versioned module packages (i.e. non-main module,
non-directory replaced), or in the case a versioned target does not
exist in the gobin cache.