main: only use local cache on install for -nonet
Using the local module cache for installs unconditionally means that
there will be situations where that cache is incomplete and hence the
install (and any subsequent step) will fail.

Rather we only need to enforce the use of the local module cache when we
are in -nonet.

Hence we now allow network access in the case of a module cache miss in
the non-nonet situation, which is exactly the semantics we intend.

Test to follow.

Fixes #86